:memo: Fix: UI elements
