+++ PoliFitnessApp/app/src/main/java/com/uca/polifitnessapp/ui/contactscreen/ui/ContactScreen.kt
@@ -42,11 +42,12 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
+import com.uca.polifitnessapp.ui.navigation.components.BackButton
 import com.uca.polifitnessapp.ui.theme.md_theme_light_outline
 
 @Composable
 fun Contact(
-    navController: NavController
+    onBackPress: () -> Unit,
 ) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
@@ -58,7 +59,7 @@ fun Contact(
         BackButton(
             modifier = Modifier
                 .align(Alignment.Start),
-            navController
+            onBackPress
         )
         Spacer(modifier = Modifier.height(25.dp))
 
@@ -74,31 +75,6 @@ fun Contact(
         SendEmail()
     }
 }
-
-@Composable
-fun BackButton(
-    modifier: Modifier = Modifier,
-    navController: NavController
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        // center items horizontally in the row
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Start
-    ) {
-        IconButton(
-            onClick = { navController.popBackStack() }
-        ) {
-            Icon(
-                Icons.Outlined.ArrowBack,
-                contentDescription = "Back button"
-            )
-        }
-    }
-}
-
 
 @Composable
 fun ContactTitle() {
