+++ PoliFitnessApp/app/src/main/java/com/uca/polifitnessapp/ui/user/ui/EditProfileScreen.kt
@@ -49,10 +49,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.uca.polifitnessapp.R
+import com.uca.polifitnessapp.ui.navigation.components.BackButton
 import com.uca.polifitnessapp.ui.user.viewmodel.EditProfileViewModel
 import com.uca.polifitnessapp.ui.user.viewmodel.UserViewModel
-
-
 
 
 @Composable
@@ -513,33 +512,6 @@ fun SaveButton(
             fontSize = 16.sp,
             color = Color.White,
             fontWeight = FontWeight.Bold
-        )
-    }
-}
-
-@Composable
-fun BackButton(
-    modifier: Modifier = Modifier,
-    onBackPress : () -> Unit
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth(),
-        // center items horizontally in the row
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Start
-    ) {
-        IconButton(
-            onClick = { onBackPress() }
-        ) {
-            Icon(
-                Icons.Outlined.ArrowBack,
-                contentDescription = "Back button"
-            )
-        }
-        Text(
-            text = "Regresar",
-            style = MaterialTheme.typography.titleSmall
         )
     }
 }
